+++ packages/story-editor/src/app/rightClickMenu/provider.js
@@ -40,11 +40,11 @@ import { v4 as uuidv4 } from 'uuid';
 import { useStory } from '..';
 import { createPage, duplicatePage, ELEMENT_TYPES } from '../../elements';
 import updateProperties from '../../components/inspector/design/updateProperties';
-import useAddPreset from '../../components/panels/design/preset/useAddPreset';
+import useAddPreset from '../../utils/useAddPreset';
 import useApplyStyle from '../../components/panels/design/preset/stylePreset/useApplyStyle';
-import { PRESET_TYPES } from '../../components/panels/design/preset/constants';
+import { PRESET_TYPES } from '../../constants';
 import { useCanvas } from '../canvas';
-import { getTextPresets } from '../../components/panels/design/preset/utils';
+import { getTextPresets } from '../../utils/presetUtils';
 import getUpdatedSizeAndPosition from '../../utils/getUpdatedSizeAndPosition';
 import { useHistory } from '../history';
 import useDeletePreset from '../../components/panels/design/preset/useDeletePreset';
